+++ ChangeLog.md
@@ -1,6 +1,10 @@
 # IPFS-Toolkit Progressive ChangeLog 
 This library is still under development and is currently being tested in various use-case projects. Due to its early stage of development, many successive versions of this library are not fully backward-compatible with their previous versions.
 
+## v0.2.5:
+ipfshttpclient2: bugfix correcting unintended IDE import rearrangement
+IPFS_DataTransmission: made encryption & decryption callbacks private attributes 
+
 ## v0.2.4:
 IPFS_API:
   - SubscribeToTopic(): Returns a listener object (PubsubListener), on which Terminate() and Listen() functions can be called to stop and restart the PubSub Subscription.